Setup jest & supertest, split Express app, add test environment
