+++ server/test/synths.test.js
@@ -1,0 +1,11 @@
+const app = require('../app');
+
+const request = require('supertest');
+
+test('should try', async (done) => {
+  const what = await request(app).get('/synths');
+  //   console.log(what);
+
+  expect(what).toBeDefined();
+  done();
+});
